modify int to double for relic items

+++ src/main/java/com/guildedrose/balance/BalanceViewer.java
@@ -1,5 +1,5 @@
 package com.guildedrose.balance;
 
 public interface BalanceViewer {
-    int getBalance();
+    double getBalance();
 }
